api/user/info added

+++ src/models/product.ts
@@ -27,7 +27,7 @@ Product.init(
       allowNull: false,
     },
     price: {
-      type: DataTypes.FLOAT(10, 2),
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
     },
     image_url_small: {
